+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 12.67.0
 
 * `[usage]` Added support of raw version printing
+* `[path]` Added method `JoinSecure` - more "secure" alternative to standard `Join`
 * `[usage]` Code refactoring
 
 ### 12.66.0